refactor: share the composable UI-binding helper
Extract two jscpd-flagged composable clones into named shared helpers:
- fetchJson (src/utils/fetchJson.ts): normalise a fetch into a
  { ok, data } | { ok:false, error, status } union. accountingUi.apiCall
  and collectionUi.apiGet/apiSend now delegate to it (the two package
  result types ApiResult / CollectionApiResult are structurally identical).
- useAutoRefresh (src/composables/useAutoRefresh.ts): refresh on mount,
  window focus, and watched-source change. useHeaderButtons and
  useSessionContext now call it instead of hand-wiring the lifecycle.

Pure refactor; behavior identical.
